requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
behave==1.2.6
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
EasyProcess==1.1
enchant==0.0.1
entrypoint2==1.1
exceptiongroup==1.1.3
execnet==2.0.2
fasteners==0.19
filelock==3.13.1
fonttools==4.44.1
h11==0.14.0
ibm-cos-sdk==2.13.2
ibm-cos-sdk-core==2.13.2
ibm-cos-sdk-s3transfer==2.13.2
ibm-watson-machine-learning==1.0.327
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
jmespath==1.0.1
kiwisolver==1.4.5
lomond==0.3.3
markdown-it-py==3.0.0
matplotlib==3.8.1
mdurl==0.1.2
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
parameterized==0.9.0
parse==1.19.1
parse-type==0.6.2
pdbp==1.5.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
py==1.11.0
PyAutoGUI==0.9.54
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pynose==1.4.8
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyotp==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==7.4.3
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
requests==2.31.0
rich==13.7.0
rubicon-objc==0.4.7
sbvirtualdisplay==1.3.0
selenium==4.15.2
seleniumbase==4.21.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.0
tabulate==0.9.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==1.26.18
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.17.0
